feat: merged camera systems
